requirements.txt
opencv-python
scikit-image
numpy
Pillow
tqdm
imagehash
face_recognition
dlib
pandas


